pyproject.toml
[project]
name = "transfermaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests",
    "Pillow",
    "pillow-heif",
    "vtracer",
    "cairosvg",
    "shapely",
    "lxml",
]
